astro-panchang/requirements.txt
pytz
certifi
pyswisseph
geopy
timezonefinder